back/requirements.txt
# Core
Django==5.2.5
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.7.0
django-storages==1.14.6
drf-spectacular==0.28.0
asgiref==3.9.1
sqlparse==0.5.3
pytz==2025.2
tzdata==2025.2

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.43

# Auth & Security
cryptography==45.0.6
PyJWT==2.10.1

# HTTP & Networking
requests==2.32.5
requests-toolbelt==1.0.0
httpx==0.28.1
httpcore==1.0.9
urllib3==2.5.0
idna==3.10
certifi==2025.8.3
sniffio==1.3.1
anyio==4.11.0
h11==0.16.0

# LangChain / AI
langchain==0.3.27
langchain-core==0.3.77
langchain-text-splitters==0.3.11
langsmith==0.4.31
openai==2.0.1
transformers==4.35.2
torch==2.5.1
accelerate==0.24.1
sentencepiece==0.1.96
protobuf==3.20.3   

# Utils
attrs==25.3.0
packaging==25.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.1
greenlet==3.2.4
zstandard==0.25.0
orjson==3.11.3
PyYAML==6.0.3
python-decouple==3.8
python-dotenv==1.1.1
inflection==0.5.1

# JSON Schema
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonpatch==1.33
jsonpointer==3.0.0
referencing==0.36.2
rpds-py==0.27.1

# Extra
cffi==1.17.1
pycparser==2.23
annotated-types==0.7.0
jiter==0.11.0
distro==1.9.0
tenacity==9.1.2
